[oidc] Export getTokenPayload; use it in functions image optimization
Move the unverified JWT payload decode from token-util into a
runtime-agnostic token-payload module (Buffer with atob fallback),
export it from the node, edge-light, and browser entrypoints, and type
the Vercel claims (project_id, owner_id, environment). The functions
image-optimization module now delegates claim decoding to @vercel/oidc
instead of hand-rolling base64url parsing.
