+++ .changeset/oidc-token-payload-export.md
@@ -1,0 +1,5 @@
+---
+'@vercel/oidc': minor
+---
+
+Export `getTokenPayload` and the `TokenPayload` type for decoding a Vercel OIDC token's claims (without signature verification) from all runtime entrypoints.
